+++ ucloud/resource_ucloud_instance.go
@@ -120,6 +120,7 @@ func resourceUCloudInstance() *schema.Resource {
 					"local_ssd",
 					"cloud_normal",
 					"cloud_ssd",
+					"cloud_rssd",
 				}, false),
 			},
 
@@ -142,6 +143,38 @@ func resourceUCloudInstance() *schema.Resource {
 					"local_normal",
 					"local_ssd",
 				}, false),
+			},
+
+			"data_disks": {
+				Type:     schema.TypeList,
+				Optional: true,
+				MinItems: 1,
+				MaxItems: 1,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"size": {
+							Type:     schema.TypeInt,
+							Required: true,
+							ForceNew: true,
+						},
+						"type": {
+							Type:     schema.TypeString,
+							Required: true,
+							ValidateFunc: validation.StringInSlice([]string{
+								"cloud_normal",
+								"cloud_ssd",
+								"cloud_rssd",
+							}, false),
+							ForceNew: true,
+						},
+					},
+				},
+			},
+
+			"delete_disks_with_instance": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  true,
 			},
 
 			"remark": {
@@ -290,7 +323,7 @@ func resourceUCloudInstanceCreate(d *schema.ResourceData, meta interface{}) erro
 	var bootDiskType string
 	if v, ok := d.GetOk("boot_disk_type"); ok {
 		if v == "cloud_normal" {
-			choices := []string{"local_normal", "local_ssd", "cloud_ssd"}
+			choices := []string{"local_normal", "local_ssd", "cloud_ssd", "cloud_rssd"}
 			return fmt.Errorf("the %q of boot disk type is not supported currently, please try one of %v", "cloud_normal", choices)
 		}
 		bootDiskType = v.(string)
@@ -360,7 +393,7 @@ func resourceUCloudInstanceCreate(d *schema.ResourceData, meta interface{}) erro
 			return fmt.Errorf("expected boot_disk_size to be at least %d", imageResp.ImageSize)
 		}
 
-		if bootDiskType == "cloud_normal" || bootDiskType == "cloud_ssd" {
+		if bootDiskType == "cloud_normal" || bootDiskType == "cloud_ssd" || bootDiskType == "cloud_rssd" {
 			bootDisk.Size = ucloud.Int(v.(int))
 		}
 	}
@@ -382,6 +415,17 @@ func resourceUCloudInstanceCreate(d *schema.ResourceData, meta interface{}) erro
 		dataDisk.Size = ucloud.Int(v.(int))
 
 		req.Disks = append(req.Disks, dataDisk)
+	}
+
+	if v, ok := d.GetOk("data_disks"); ok {
+		for _, item := range v.([]interface{}) {
+			dataDisk := uhost.UHostDisk{}
+			dataDisk.IsBoot = ucloud.String("false")
+			disk := item.(map[string]interface{})
+			dataDisk.Size = ucloud.Int(disk["size"].(int))
+			dataDisk.Type = ucloud.String(upperCvt.unconvert(disk["type"].(string)))
+			req.Disks = append(req.Disks, dataDisk)
+		}
 	}
 
 	// if tag is empty string, use default tag
@@ -561,7 +605,7 @@ func resourceUCloudInstanceUpdate(d *schema.ResourceData, meta interface{}) erro
 		}
 
 		// the initialization of cloud boot disk is done at creation instance
-		if ((bootDiskType == "cloud_normal" || bootDiskType == "cloud_ssd") && d.IsNewResource()) || imageResp.ImageSize == bootDiskSize {
+		if ((bootDiskType == "cloud_normal" || bootDiskType == "cloud_ssd" || bootDiskType == "cloud_rssd") && d.IsNewResource()) || imageResp.ImageSize == bootDiskSize {
 			bootDiskNeedUpdate = false
 		} else {
 			bootDiskReq.DiskSpace = ucloud.Int(bootDiskSize)
@@ -882,8 +926,7 @@ func resourceUCloudInstanceDelete(d *schema.ResourceData, meta interface{}) erro
 
 	deleReq := conn.NewTerminateUHostInstanceRequest()
 	deleReq.UHostId = ucloud.String(d.Id())
-	deleReq.ReleaseUDisk = ucloud.Bool(true)
-	deleReq.ReleaseEIP = ucloud.Bool(false)
+	deleReq.ReleaseUDisk = ucloud.Bool(d.Get("delete_disks_with_instance").(bool))
 
 	return resource.Retry(15*time.Minute, func() *resource.RetryError {
 		instance, err := client.describeInstanceById(d.Id())
@@ -1052,7 +1095,7 @@ func diffValidateBootDiskTypeWithDataDiskType(diff *schema.ResourceDiff, meta in
 			return fmt.Errorf("the data_disk_type %q must be same as boot_disk_type %q", dataDiskType, bootDiskType)
 		}
 	}
-	if checkStringIn(bootDiskType, []string{"cloud_normal", "cloud_ssd"}) == nil && checkStringIn(dataDiskType, []string{"local_normal", "local_ssd"}) == nil {
+	if checkStringIn(bootDiskType, []string{"cloud_normal", "cloud_ssd", "cloud_rssd"}) == nil && checkStringIn(dataDiskType, []string{"local_normal", "local_ssd"}) == nil {
 		return fmt.Errorf("the instance cannot have local data disk, When the %q is %q", "boot_disk_type", bootDiskType)
 	}